refactor(utils): extract ANSI colour helpers into mtui/colors.py
Move green/red/yellow/blue verbatim out of mtui/utils.py into a new
mtui/colors.py module. utils.py keeps a re-export shim for callers so
this commit is byte-identical at the public-API surface.

First commit of the Phase 5b / C8 split (PLAN.md item 11). The shim
goes away in a later commit once all in-tree call sites are rewritten.
